feat: membership - membership get detail current year api#94
- season Integer로 타입 변경
- read current season api 추가

+++ membership-service/src/main/java/com/boeingmerryho/business/membershipservice/domain/entity/MembershipUser.java
@@ -34,7 +34,7 @@ public class MembershipUser extends BaseEntity {
 	private Long userId;
 
 	@Column(nullable = false)
-	private Long season;
+	private Integer season;
 
 	@Column(nullable = false)
 	private Boolean isActive;
